WO: Broke the map manager up into parital class (keep the logic sperate from networking implementation).
