+++ Assets/Scripts/Managers/TTT.Managers.asmdef
@@ -8,7 +8,8 @@
         "GUID:84651a3751eca9349aac36a66bba901b",
         "GUID:4c861275823ef884d91df8647f4caa34",
         "GUID:0584c7e699b26eb43bd6f8f5022c28bc",
-        "GUID:840341cf92912f048a97909ccc73bc49"
+        "GUID:840341cf92912f048a97909ccc73bc49",
+        "GUID:91ba5b220a43d5443ae04d679286f57d"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
